WiFi Autoconfig and temp reading using TMP102
This is the current state of the project that uses my old TMP102 sensor
over I2C and forwards the readings and node name using dweet.io.

The WiFi autoconfigure is pretty complete and allows Wifi credentials
and node name to be entered when the node is in a vanilla state.
The node can be reset to vanilla using GPIO-8 and the node can be forced
to read the temp using GPIO-7.